+++ src/effect/modules/products/import/utils.ts
@@ -1,8 +1,22 @@
+import { Effect } from 'effect';
 import { parse } from 'csv-parse/sync';
+import { LocationType } from '@stocket/types/locations';
+import type { PhotosService } from '../../photos/service';
+import {
+  ProductImportCsvParseFailed,
+  ProductImportUnsupportedFormat,
+  ProductsInfrastructureError,
+} from '../products.errors';
+import type { ProductImportRepository } from './repository';
 import type {
   CsvParseResult,
   CsvRecord,
+  ImportAreaRow,
+  ImportCaches,
+  ImportCategoryRow,
+  ImportLocationRow,
   ImportProductRow,
+  ImportProductsFromCsvOptions,
   NormalizedProductImportRow,
   ProductImportCommitResultDto,
   ProductImportErrorDto,
@@ -43,6 +57,31 @@ export const makeEmptyProductImportCommitResult =
     photosImported: 0,
     warnings: [],
   });
+
+export function parseProductImportRequest(
+  content: string,
+  importType: ImportProductsFromCsvOptions['importType'],
+) {
+  return Effect.gen(function* () {
+    const parsed = yield* Effect.try({
+      try: () => parseCsvContent(content),
+      catch: (cause) =>
+        new ProductImportCsvParseFailed({
+          cause,
+          messageKey: 'products.importCsvParseFailed',
+        }),
+    });
+    const format = detectProductImportFormat(parsed.headers, importType);
+    if (!format) {
+      return yield* Effect.fail(
+        new ProductImportUnsupportedFormat({
+          messageKey: 'products.importUnsupportedFormat',
+        }),
+      );
+    }
+    return { parsed, format };
+  });
+}
 
 const hasAllHeaders = (
   headerSet: ReadonlySet<string>,
@@ -499,6 +538,476 @@ export function makeProductImportPreview(
     issues,
     warnings,
   };
+}
+
+const getOrCreateCategoryPath = (
+  repository: ProductImportRepository,
+  categoryPath: string,
+  caches: ImportCaches,
+  result: ProductImportResultDto,
+) =>
+  Effect.gen(function* () {
+    const parts = normalizeCategoryPath(categoryPath)
+      .split('/')
+      .map((part) => part.trim())
+      .filter(Boolean);
+
+    if (parts.length === 0) {
+      return yield* Effect.fail(
+        new ProductsInfrastructureError({
+          action: 'resolve import category path',
+          messageKey: 'products.repositoryFailed',
+        }),
+      );
+    }
+
+    let parentId: string | null = null;
+    let categoryId = '';
+
+    for (const part of parts) {
+      const cacheKey = `${parentId ?? 'root'}:${part}`;
+      const cached = caches.categories.get(cacheKey);
+      if (cached) {
+        parentId = cached;
+        categoryId = cached;
+        continue;
+      }
+
+      let category: ImportCategoryRow | null =
+        yield* repository.findCategoryByNameAndParent(part, parentId);
+
+      if (!category) {
+        category = yield* repository.createCategory({
+          name: part,
+          parent_id: parentId,
+          description: 'Imported via product import',
+        });
+        result.categoriesCreated++;
+      }
+
+      caches.categories.set(cacheKey, category.id);
+      parentId = category.id;
+      categoryId = category.id;
+    }
+
+    return categoryId;
+  });
+
+const getOrCreateLocation = (
+  repository: ProductImportRepository,
+  locationName: string,
+  caches: ImportCaches,
+  result: ProductImportCommitResultDto,
+) =>
+  Effect.gen(function* () {
+    const name = locationName.trim();
+    if (name === '') return null;
+
+    const cached = caches.locations.get(name);
+    if (cached) return cached;
+
+    let location: ImportLocationRow | null =
+      yield* repository.findLocationByName(name);
+
+    if (!location) {
+      location = yield* repository.createLocation({
+        name,
+        type: LocationType.WAREHOUSE,
+        address: '',
+        contact_person: '',
+        phone: '',
+        is_active: true,
+      });
+      result.locationsCreated++;
+    }
+
+    caches.locations.set(name, location.id);
+    return location.id;
+  });
+
+const getOrCreateAreaPath = (
+  repository: ProductImportRepository,
+  locationId: string,
+  areaPath: string,
+  caches: ImportCaches,
+  result: ProductImportCommitResultDto,
+) =>
+  Effect.gen(function* () {
+    const parts = areaPath
+      .split('/')
+      .map((part) => part.trim())
+      .filter(Boolean);
+    if (parts.length === 0) return null;
+
+    let parentId: string | null = null;
+    let areaId = '';
+
+    for (const part of parts) {
+      const cacheKey = `${locationId}:${parentId ?? 'root'}:${part}`;
+      const cached = caches.areas.get(cacheKey);
+      if (cached) {
+        parentId = cached;
+        areaId = cached;
+        continue;
+      }
+
+      let area: ImportAreaRow | null =
+        yield* repository.findAreaByNameLocationAndParent(
+          part,
+          locationId,
+          parentId,
+        );
+
+      if (!area) {
+        area = yield* repository.createArea({
+          location_id: locationId,
+          parent_id: parentId,
+          name: part,
+          code: '',
+          description: 'Imported via product import',
+          is_active: true,
+        });
+        result.areasCreated++;
+      }
+
+      caches.areas.set(cacheKey, area.id);
+      parentId = area.id;
+      areaId = area.id;
+    }
+
+    return areaId;
+  });
+
+const upsertProduct = (
+  repository: ProductImportRepository,
+  row: NormalizedProductImportRow,
+  categoryId: string,
+  caches: ImportCaches,
+  result: ProductImportResultDto,
+  expiryDate: Date | null,
+  userId: string,
+) =>
+  Effect.gen(function* () {
+    const updatedBy = userId;
+    const cached = caches.products.get(row.sku);
+    if (cached) return cached;
+
+    const values: ProductImportValues = {
+      name: row.name,
+      description: nullableText(row.description),
+      category_id: categoryId,
+      unit: nullableText(row.unit),
+      barcode: nullableText(row.barcode),
+      standard_price: parseProductImportNumber(row.standard_price),
+      reorder_point: parseInteger(row.reorder_point, 0),
+      is_active: parseBoolean(row.is_active, true),
+      is_perishable: parseBoolean(row.is_perishable, Boolean(expiryDate)),
+      notes: nullableText(row.notes),
+    };
+
+    const existing = yield* repository.findProductBySku(row.sku);
+    if (!existing) {
+      const product = yield* repository.createProduct({
+        sku: row.sku,
+        ...values,
+        created_by: updatedBy,
+        updated_by: updatedBy,
+      });
+      result.productsCreated++;
+      caches.products.set(row.sku, product);
+      return product;
+    }
+
+    if (productValuesMatch(existing, values)) {
+      caches.products.set(row.sku, existing);
+      return existing;
+    }
+
+    const product = yield* repository.updateProduct(existing.id, {
+      ...values,
+      updated_by: updatedBy,
+    });
+    if (!product) {
+      return yield* Effect.fail(
+        new ProductsInfrastructureError({
+          action: 'update import product',
+          messageKey: 'products.repositoryFailed',
+        }),
+      );
+    }
+    result.productsUpdated++;
+    caches.products.set(row.sku, product);
+    return product;
+  });
+
+const upsertInventory = (
+  repository: ProductImportRepository,
+  product: ImportProductRow,
+  locationId: string | null,
+  areaId: string | null,
+  row: NormalizedProductImportRow,
+  result: ProductImportCommitResultDto,
+  expiryDate: Date | null,
+) =>
+  Effect.gen(function* () {
+    if (!locationId) return;
+
+    const existing = yield* repository.findInventoryByProductLocationArea(
+      product.id,
+      locationId,
+      areaId,
+    );
+    const quantity = parseInteger(row.quantity, 0);
+    if (areaId === null) {
+      const hasAreaScopedInventory =
+        yield* repository.hasAreaScopedInventoryForProductAndLocation(
+          product.id,
+          locationId,
+        );
+      if (hasAreaScopedInventory) {
+        return yield* Effect.fail(
+          new ProductsInfrastructureError({
+            action: 'import root inventory with area-scoped inventory',
+            messageKey: 'products.importAreaScopedInventoryConflict',
+          }),
+        );
+      }
+    }
+
+    if (!existing) {
+      yield* repository.createInventory({
+        product_id: product.id,
+        location_id: locationId,
+        area_id: areaId,
+        quantity,
+        expiry_date: expiryDate,
+      });
+      result.inventoryRecordsCreated++;
+      return;
+    }
+
+    const inventory = yield* repository.updateInventory(existing.id, {
+      quantity,
+      expiry_date: expiryDate,
+    });
+    if (!inventory) {
+      return yield* Effect.fail(
+        new ProductsInfrastructureError({
+          action: 'update import inventory',
+          messageKey: 'products.repositoryFailed',
+        }),
+      );
+    }
+    result.inventoryRecordsUpdated++;
+  });
+
+const importPhotosForProduct = (
+  photosService: PhotosService,
+  product: ImportProductRow,
+  row: NormalizedProductImportRow,
+  result: ProductImportCommitResultDto,
+  userId: string,
+  importedPhotoProducts: Set<string>,
+) =>
+  Effect.gen(function* () {
+    if (row.photo_urls.length === 0 || importedPhotoProducts.has(product.id)) {
+      return;
+    }
+    importedPhotoProducts.add(product.id);
+
+    for (const photoUrl of row.photo_urls) {
+      const response = yield* Effect.tryPromise({
+        try: () => fetch(photoUrl),
+        catch: (cause) => cause,
+      }).pipe(
+        Effect.catchAll((cause) =>
+          Effect.sync(() => {
+            pushWarning(
+              result,
+              `Failed to download Sortly photo ${photoUrl}: ${formatImportError(cause)}`,
+              row.sourceRow,
+            );
+            return null;
+          }),
+        ),
+      );
+      if (!response) continue;
+
+      if (!response.ok) {
+        pushWarning(
+          result,
+          `Failed to download Sortly photo ${photoUrl}: HTTP ${response.status}`,
+          row.sourceRow,
+        );
+        continue;
+      }
+
+      const contentType =
+        response.headers.get('content-type') ?? 'application/octet-stream';
+      const arrayBuffer = yield* Effect.tryPromise({
+        try: () => response.arrayBuffer(),
+        catch: (cause) => cause,
+      }).pipe(
+        Effect.catchAll((cause) =>
+          Effect.sync(() => {
+            pushWarning(
+              result,
+              `Failed to read Sortly photo ${photoUrl}: ${formatImportError(cause)}`,
+              row.sourceRow,
+            );
+            return null;
+          }),
+        ),
+      );
+      if (!arrayBuffer) continue;
+
+      const buffer = Buffer.from(arrayBuffer);
+      const filename = photoUrl.split('/').pop()?.split('?')[0] || 'sortly-photo';
+      yield* photosService
+        .uploadPhoto(
+          product.id,
+          {
+            originalname: filename,
+            mimetype: contentType,
+            size: buffer.length,
+            buffer,
+          },
+          userId,
+        )
+        .pipe(
+          Effect.match({
+            onFailure: (error) => {
+              pushWarning(
+                result,
+                `Failed to import Sortly photo ${photoUrl}: ${formatImportError(error)}`,
+                row.sourceRow,
+              );
+            },
+            onSuccess: () => {
+              result.photosImported++;
+            },
+          }),
+        );
+    }
+  });
+
+const validateRootInventoryImport = (
+  repository: ProductImportRepository,
+  row: NormalizedProductImportRow,
+  caches: ImportCaches,
+) =>
+  Effect.gen(function* () {
+    const locationName = row.location.trim();
+    if (locationName === '' || row.area_path.trim() !== '') return;
+
+    const cachedProduct = caches.products.get(row.sku);
+    let product = cachedProduct ?? null;
+    if (!product) {
+      product = yield* repository.findProductBySku(row.sku);
+    }
+    if (!product) return;
+
+    let locationId = caches.locations.get(locationName) ?? null;
+    if (!locationId) {
+      const location = yield* repository.findLocationByName(locationName);
+      locationId = location?.id ?? null;
+    }
+    if (!locationId) return;
+
+    const hasAreaScopedInventory =
+      yield* repository.hasAreaScopedInventoryForProductAndLocation(
+        product.id,
+        locationId,
+      );
+    if (hasAreaScopedInventory) {
+      return yield* Effect.fail(
+        new ProductsInfrastructureError({
+          action: 'import root inventory with area-scoped inventory',
+          messageKey: 'products.importAreaScopedInventoryConflict',
+        }),
+      );
+    }
+  });
+
+export function importProductImportRow({
+  repository,
+  photosService,
+  row,
+  caches,
+  result,
+  expiryDate,
+  userId,
+  importedPhotoProducts,
+  importPhotos,
+}: {
+  readonly repository: ProductImportRepository;
+  readonly photosService: PhotosService;
+  readonly row: NormalizedProductImportRow;
+  readonly caches: ImportCaches;
+  readonly result: ProductImportCommitResultDto;
+  readonly expiryDate: Date | null;
+  readonly userId: string;
+  readonly importedPhotoProducts: Set<string>;
+  readonly importPhotos: boolean;
+}) {
+  return Effect.gen(function* () {
+    if (row.area_path.trim() !== '' && row.location.trim() === '') {
+      return yield* Effect.fail(
+        new Error('Cannot import area_path without location'),
+      );
+    }
+
+    yield* validateRootInventoryImport(repository, row, caches);
+    const categoryId = yield* getOrCreateCategoryPath(
+      repository,
+      row.category_path,
+      caches,
+      result,
+    );
+    const product = yield* upsertProduct(
+      repository,
+      row,
+      categoryId,
+      caches,
+      result,
+      expiryDate,
+      userId,
+    );
+    const locationId = yield* getOrCreateLocation(
+      repository,
+      row.location,
+      caches,
+      result,
+    );
+    const areaId = locationId
+      ? yield* getOrCreateAreaPath(
+          repository,
+          locationId,
+          row.area_path,
+          caches,
+          result,
+        )
+      : null;
+    yield* upsertInventory(
+      repository,
+      product,
+      locationId,
+      areaId,
+      row,
+      result,
+      expiryDate,
+    );
+    if (importPhotos) {
+      yield* importPhotosForProduct(
+        photosService,
+        product,
+        row,
+        result,
+        userId,
+        importedPhotoProducts,
+      );
+    }
+  });
 }
 
 export function parseDate(value: string): Date | null {
